% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tabela.R
\name{count_tabela}
\alias{count_tabela}
\title{Contagem o número de tabelas da SSP-SP
Diz quantas tabelas há no site}
\usage{
count_tabela()
}
\description{
Contagem o número de tabelas da SSP-SP
Diz quantas tabelas há no site
}
