% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_violencia_mulher.R
\name{dados_violencia_mulher}
\alias{dados_violencia_mulher}
\alias{\%>\%}
\title{Obtém dados sobre ocorrências de Violência Contra as Mulheres, no Estado de São Paulo.}
\usage{
lhs \%>\% rhs
}
\value{
Retorna uma tibble.
}
\description{
Obtém dados sobre ocorrências de Violência Contra as Mulheres, no Estado de São Paulo.
}
\examples{


}
