% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_tabela.R
\name{get_all_tabela}
\alias{get_all_tabela}
\title{Extração de todas as tabelas da SSP-SP
Extrai os dados de todas as tabelas do site}
\usage{
get_all_tabela()
}
\description{
Extração de todas as tabelas da SSP-SP
Extrai os dados de todas as tabelas do site
}
