% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mes_ano.R
\name{get_mes_ano}
\alias{get_mes_ano}
\title{Mês e ano do dado da SSP-SP
Captura, limpa e separa em colunas mês e ano das tabelas}
\usage{
get_mes_ano()
}
\description{
Mês e ano do dado da SSP-SP
Captura, limpa e separa em colunas mês e ano das tabelas
}
