% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabela.R
\name{get_tabela}
\alias{get_tabela}
\title{Extração de tabelas da SSP-SP}
\usage{
get_tabela()
}
\arguments{
\item{NULL}{}
}
\value{
tibble::tibble()
}
\description{
Cria função para ser inserida dentro de map
}
