% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{violencia_contra_mulheres_sp}
\alias{violencia_contra_mulheres_sp}
\title{Dados sobre violência contra a mulher no Estado de São Paulo}
\format{
Uma tibble.
}
\source{
SSP/SP - \url{http://www.ssp.sp.gov.br/Estatistica/ViolenciaMulher.aspx}
}
\usage{
violencia_contra_mulheres_sp
}
\description{
Neste pacote, está disponível a base de dados \code{violencia_contra_mulheres_sp},
contendo os dados \href{http://www.ssp.sp.gov.br/Estatistica/ViolenciaMulher.aspx}{divulgados pela SSP-SP}.
Os dados iniciam em Janeiro de 2019.
}
\examples{
segurancaSP::violencia_contra_mulheres_sp

}
\keyword{datasets}
